server/requirements.txt
flask
flask-cors
openai
anthropic
google-genai
